# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CND Rest API",
        "version": "7.0"
    },
    "servers": [
        {
            "url": "http://ins-2023.aml.fr/api/v7",
            "description": "REST APIs CND"
        },
        {
            "url": "https://ins-demo.cnd.info/api/v7",
            "description": "Site Demo"
        },
        {
            "url": "https://ins-bs.obamalou.fr/api/v7",
            "description": "Site Demo"
        }
    ],
    "paths": {
        "/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Returns a list of contacts",
                "operationId": "getAllContacts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Contacts"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Contacts"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "get contest main indicators",
                "operationId": "dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Dashboard"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/download/{id}": {
            "get": {
                "tags": [
                    "Download"
                ],
                "summary": "Downloads a file identified by its id",
                "operationId": "downloadGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "get by Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Contacts": {
                "title": "Contacts",
                "description": "Contact Model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "id_ecole": {
                        "type": "integer"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "teacherEmail": {
                        "type": "string"
                    },
                    "schoolName": {
                        "type": "string"
                    },
                    "schoolEmail": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Dashboard": {
                "title": "Dashboard",
                "description": "Dashboard Model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nb_candidates": {
                        "type": "integer"
                    },
                    "nb_teachers": {
                        "description": "number of teachers",
                        "type": "integer"
                    },
                    "nb_individuals": {
                        "description": "number of individuals",
                        "type": "integer"
                    },
                    "nb_duets": {
                        "description": "number of duets",
                        "type": "integer"
                    },
                    "nb_groups": {
                        "description": "number of groups",
                        "type": "integer"
                    },
                    "nb_total_groups": {
                        "description": "total number of candidates in groups",
                        "type": "integer"
                    },
                    "nb_schools": {
                        "description": "number of used schools without free candidates",
                        "type": "integer"
                    },
                    "countNbCandidateMembership": {
                        "description": "number of candidates membership",
                        "type": "integer"
                    },
                    "countNbTeacherMembership": {
                        "description": "number of teachers membership",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "title": "Error",
                "properties": {
                    "Error": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "type": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Common": {
                "title": "Common",
                "properties": {
                    "Common": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "method": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "value": {
                                    "type": "string"
                                },
                                "count": {
                                    "type": "integer"
                                },
                                "type": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Header": {
                "title": "Header",
                "properties": {
                    "objectType": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "integer"
                    },
                    "statusMessage": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Response": {
                "title": "Response",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Header"
                    },
                    {
                        "$ref": "#/components/schemas/Error"
                    },
                    {
                        "$ref": "#/components/schemas/Common"
                    }
                ]
            },
...
        }
    },
...
}